Refactor user and chat API logic into service modules
Moved user and chat-related API functions from lib/auth-service.ts and lib/api-service.ts into new service modules: services/user-service.ts and services/chat-service.ts. Updated all relevant imports in components to use the new service modules. Cleaned up lib/auth-service.ts and lib/api-service.ts, and improved type definitions and documentation for dashboard and contract APIs.
