+++ components/core/chat-sidebar.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useRef, useEffect } from "react";
 import { cn } from "@/lib/utils";
-import { sendChatMessageStream } from "@/lib/api-service";
+import { sendChatMessageStream } from "@/services/chat-service";
 import { WidgetData } from "@/context/widget-context";
 import {
   ChatHeader,
