+++ components/features/users/add-user-button.tsx
@@ -2,7 +2,7 @@
 
 import { Dialog, Button, Flex, Text } from "@radix-ui/themes";
 import { useState, useEffect } from "react";
-import { createUser } from "@/lib/auth-service";
+import { createUser } from "@/services/user-service";
 import { getMenuBranches, getMenuRoles } from "@/lib/api-service";
 import type { Branch, Role, CreateUserPayload } from "@/types";
 import {
