+++ components/features/users/create-user-dialog.tsx
@@ -20,7 +20,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Users, FileText, Mail, Lock, Briefcase } from "lucide-react";
-import { createUser } from "@/lib/auth-service";
+import { createUser } from "@/services/user-service";
 import { getMenuBranches } from "@/lib/api-service";
 import { showCreateSuccess, showError, showWarning } from "@/lib/sweetalert";
 import type { User, Branch } from "@/types/auth";
