+++ components/features/users/edit-user-dialog.tsx
@@ -20,7 +20,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Users, FileText, Mail, Briefcase } from "lucide-react";
-import { updateUser } from "@/lib/auth-service";
+import { updateUser } from "@/services/user-service";
 import { getMenuBranches } from "@/lib/api-service";
 import { showUpdateSuccess, showError, showWarning } from "@/lib/sweetalert";
 import type { User, Branch } from "@/types/auth";
